DCMAW-5594: Move BE and FE reqs into testSteps
