+++ deploy/scripts/src/mobile/testSteps/backend.ts
@@ -1,9 +1,9 @@
 import http from 'k6/http'
 import { group } from 'k6'
-import { uuidv4 } from './k6/k6-utils'
-import { getBackendUrl } from './url'
-import { parseTestClientResponse, postTestClientStart } from './test-client'
-import { isStatusCode200, isStatusCode201 } from './assertions'
+import { uuidv4 } from '../utils/k6/k6-utils'
+import { getBackendUrl } from '../utils/url'
+import { parseTestClientResponse, postTestClientStart } from '../utils/test-client'
+import { isStatusCode200, isStatusCode201 } from '../utils/assertions'
 
 export function postVerifyAuthorizeRequest (): string {
   let verifyUrl: string
